+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.4.0] - 2022-11-03
+
 ### Incompatible changes
 
 - More string and number values are exported correctly:
@@ -108,7 +110,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added the full feature as a library that can be embedded into other apps.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/MForster/factorio-export/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/MForster/factorio-export/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/MForster/factorio-export/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/MForster/factorio-export/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/MForster/factorio-export/compare/v0.1.2...v0.2.0
 [0.1.2]: https://github.com/MForster/factorio-export/compare/v0.1.1...v0.1.2